refactor(gateway): extract bounded node invocation policies
