+++ src/gateway/server-methods/nodes.invoke-deadline.ts
@@ -1,0 +1,44 @@
+import { MAX_TIMER_TIMEOUT_MS } from "@openclaw/normalization-core/number-coercion";
+
+export const NODE_INVOKE_DEADLINE_EXPIRED = Symbol("node invoke deadline expired");
+
+/** Bounds node pairing, wake, policy, and transport preparation by one absolute deadline. */
+export async function awaitNodeInvokeWithinDeadline<T>(
+  operation: () => Promise<T>,
+  deadlineAtMs: number | undefined,
+): Promise<T | typeof NODE_INVOKE_DEADLINE_EXPIRED> {
+  if (deadlineAtMs === undefined) {
+    return await operation();
+  }
+  if (Math.max(0, deadlineAtMs - Date.now()) === 0) {
+    return NODE_INVOKE_DEADLINE_EXPIRED;
+  }
+
+  let timer: ReturnType<typeof setTimeout> | undefined;
+  try {
+    // Arm the timer before plugin or push code can synchronously consume the
+    // budget; timer callbacks alone cannot establish an absolute deadline.
+    const deadline = new Promise<typeof NODE_INVOKE_DEADLINE_EXPIRED>((resolve) => {
+      // Node overflows long timer delays, so rearm against the real deadline.
+      const waitForDeadline = () => {
+        const remainingMs = Math.max(0, deadlineAtMs - Date.now());
+        if (remainingMs === 0) {
+          resolve(NODE_INVOKE_DEADLINE_EXPIRED);
+          return;
+        }
+        timer = setTimeout(waitForDeadline, Math.min(remainingMs, MAX_TIMER_TIMEOUT_MS));
+      };
+      waitForDeadline();
+    });
+    return await Promise.race([
+      deadline,
+      operation().then((result) =>
+        Date.now() >= deadlineAtMs ? NODE_INVOKE_DEADLINE_EXPIRED : result,
+      ),
+    ]);
+  } finally {
+    if (timer !== undefined) {
+      clearTimeout(timer);
+    }
+  }
+}
