+++ src/gateway/server-methods/nodes.invoke-foreground.ts
@@ -1,0 +1,32 @@
+import {
+  normalizeLowercaseStringOrEmpty,
+  normalizeOptionalString,
+} from "@openclaw/normalization-core/string-coerce";
+import { isForegroundRestrictedPluginNodeCommand } from "../node-command-policy.js";
+
+/** Queues only commands that iOS explicitly rejected as requiring the foreground. */
+export function shouldQueueAsPendingForegroundAction(params: {
+  platform?: string;
+  command: string;
+  error: unknown;
+}): boolean {
+  const platform = normalizeLowercaseStringOrEmpty(params.platform);
+  if (!platform.startsWith("ios") && !platform.startsWith("ipados")) {
+    return false;
+  }
+  if (
+    !isForegroundRestrictedPluginNodeCommand(params.command) &&
+    !params.command.startsWith("camera.") &&
+    !params.command.startsWith("screen.") &&
+    !params.command.startsWith("talk.")
+  ) {
+    return false;
+  }
+  const error =
+    params.error && typeof params.error === "object"
+      ? (params.error as { code?: unknown; message?: unknown })
+      : null;
+  const code = normalizeOptionalString(error?.code)?.toUpperCase() ?? "";
+  const message = normalizeOptionalString(error?.message)?.toUpperCase() ?? "";
+  return code === "NODE_BACKGROUND_UNAVAILABLE" || message.includes("BACKGROUND_UNAVAILABLE");
+}
